POI：move duplicated NullOutputStream classes to test suite utilities
