+++ testcases/main/NPOI.TestCases.csproj
@@ -254,6 +254,7 @@
     <Compile Include="SS\Util\TestSheetUtil.cs" />
     <Compile Include="TestPOIDocumentMain.cs" />
     <Compile Include="TestPOITestCase.cs" />
+    <Compile Include="Util\NullOutputStream.cs" />
     <Compile Include="Util\TestBitField.cs" />
     <Compile Include="Util\TestByteField.cs" />
     <Compile Include="Util\TestHexDump.cs" />
