+++ testcases/main/SS/UserModel/BaseTestWorkbook.cs
@@ -28,6 +28,7 @@ namespace TestCases.SS.UserModel
     using NPOI.Util;
     using NPOI.SS;
     using TestCases.HSSF;
+    using TestCases.Util;
 
     /**
      * @author Yegor Kozlov
@@ -829,46 +830,7 @@ namespace TestCases.SS.UserModel
         }
 
 
-        protected class NullOutputStream : OutputStream
-        {
-            public NullOutputStream()
-            {
-            }
-
-            public override bool CanRead => throw new NotImplementedException();
-
-            public override bool CanSeek => throw new NotImplementedException();
-
-            public override bool CanWrite => throw new NotImplementedException();
-
-            public override long Length => throw new NotImplementedException();
-
-            public override long Position { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
-
-            public override void Flush()
-            {
-                throw new NotImplementedException();
-            }
-
-            public override int Read(byte[] buffer, int offset, int count)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override long Seek(long offset, SeekOrigin origin)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override void SetLength(long value)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override void Write(int b)
-            {
-            }
-        }
+        
 
         [Test]
         public void Test58499()
